object_detection/requirements.txt
numpy
pandas
tensorflow
Pillow
matplotlib
opencv-python
object_detection
pyqt
